requirements.txt
discord.py[voice]==2.3.1
python-dotenv==1.0.0
pytube==15.0.0
spotipy==2.23.0
kthread